support assignments for MUL, DIV, MOD
