+++ README.md
@@ -473,6 +473,9 @@ BC = ($C000)  ; expand to -> LD BC, ($C000)
 A += B        ; expand to -> ADD A, B
 HL += DE      ; expand to -> ADD HL, DE
 A -= B        ; expand to -> SUB A, B
+A *= B        ; expand to -> MUL A, B
+A /= B        ; expand to -> DIV A, B
+A %= B        ; expand to -> MOD A, B
 A &= B        ; expand to -> AND A, B
 A |= B        ; expand to -> OR A, B
 A ^= B        ; expand to -> XOR A, B
