+++ src/file.hpp
@@ -250,12 +250,37 @@ LineData::LineData(const char* path, int lineNumber, std::string text)
                     }
                     break;
                 case '/':
-                    cp++;
-                    this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Div, "/"));
+                    if (*(cp + 1) == '=') {
+                        cp += 2;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::EqualDiv, "/="));
+                    } else {
+                        cp++;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Div, "/"));
+                    }
                     break;
                 case '*':
-                    cp++;
-                    this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Mul, "*"));
+                    if (*(cp + 1) == '=') {
+                        cp += 2;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::EqualMul, "*="));
+                    } else {
+                        cp++;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Mul, "*"));
+                    }
+                    break;
+                case '%':
+                    if (*(cp + 1) == '=') {
+                        cp += 2;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::EqualMod, "%="));
+                    } else if (isdigit(*(cp + 1))) {
+                        for (ed = cp + 2; isdigit(*ed); ed++) {}
+                        std::string str = cp;
+                        puts((str.substr(0, ed - cp)).c_str());
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Other, str.substr(0, ed - cp)));
+                        cp = ed;
+                    } else {
+                        cp++;
+                        this->token.push_back(std::make_pair<TokenType, std::string>(TokenType::Modulo, "%"));
+                    }
                     break;
                 case '&':
                     if ('=' == *(cp + 1)) {
@@ -366,6 +391,7 @@ LineData::LineData(const char* path, int lineNumber, std::string text)
                             '^' == *ed ||
                             '<' == *ed ||
                             '>' == *ed ||
+                            '%' == *ed ||
                             ',' == *ed) {
                             break;
                         }
