+++ src/formulas.hpp
@@ -188,134 +188,3 @@ void evaluate_formulas_array(LineData* line)
         }
     }
 }
-
-void replace_assignment(LineData* line)
-{
-    std::vector<std::pair<TokenType, std::string>> left;
-    std::vector<std::pair<TokenType, std::string>> right;
-    bool isLeft = true;
-    TokenType assignment = TokenType::None;
-    for (auto it = line->token.begin(); it != line->token.end(); it++) {
-        if (it->first == TokenType::Equal ||
-            it->first == TokenType::EqualAnd ||
-            it->first == TokenType::EqualMinus ||
-            it->first == TokenType::EqualOr ||
-            it->first == TokenType::EqualPlus ||
-            it->first == TokenType::EqualXor ||
-            it->first == TokenType::EqualShiftLeft ||
-            it->first == TokenType::EqualShiftRight) {
-            if (assignment != TokenType::None) {
-                line->error = true;
-                line->errmsg = "Multiple assignments are described.";
-                return;
-            }
-            assignment = it->first;
-            isLeft = false;
-            continue;
-        }
-        if (isLeft) {
-            left.push_back(std::make_pair(it->first, it->second));
-        } else {
-            right.push_back(std::make_pair(it->first, it->second));
-        }
-    }
-    if (assignment == TokenType::None) {
-        return;
-    }
-    if (left.empty() || right.empty()) {
-        line->error = true;
-        line->errmsg = "Illegal assignment expression without left or right side.";
-        return;
-    }
-    line->token.clear();
-    std::pair<TokenType, std::string> mne;
-    switch (assignment) {
-        case TokenType::Equal:
-            mne = std::make_pair(TokenType::Other, "LD");
-            break;
-        case TokenType::EqualAnd:
-            mne = std::make_pair(TokenType::Other, "AND");
-            break;
-        case TokenType::EqualMinus:
-            mne = std::make_pair(TokenType::Other, "SUB");
-            break;
-        case TokenType::EqualOr:
-            mne = std::make_pair(TokenType::Other, "OR");
-            break;
-        case TokenType::EqualPlus:
-            mne = std::make_pair(TokenType::Other, "ADD");
-            break;
-        case TokenType::EqualXor:
-            mne = std::make_pair(TokenType::Other, "XOR");
-            break;
-        case TokenType::EqualShiftLeft:
-            mne = std::make_pair(TokenType::Other, "SL");
-            break;
-        case TokenType::EqualShiftRight:
-            mne = std::make_pair(TokenType::Other, "SR");
-            break;
-        default:
-            puts("logic error");
-            exit(-1);
-    }
-    line->token.insert(line->token.begin(), mne);
-    line->token.insert(line->token.end(), left.begin(), left.end());
-    line->token.insert(line->token.end(), std::make_pair<TokenType, std::string>(TokenType::Split, ","));
-    line->token.insert(line->token.end(), right.begin(), right.end());
-    line->isAssignmnt = true;
-}
-
-std::string oct2dec(const char* oct)
-{
-    int result = 0;
-    int len = 0;
-    while (*oct) {
-        if (isdigit(*oct) && *oct < '8') {
-            result <<= 3;
-            result += (*oct) - '0';
-            len++;
-        } else {
-            return "";
-        }
-        oct++;
-    }
-    return 0 < len ? std::to_string(result) : "";
-}
-
-std::string hex2dec(const char* hex)
-{
-    int result = 0;
-    int len = 0;
-    while (*hex) {
-        if (isdigit(*hex)) {
-            result <<= 4;
-            result += (*hex) - '0';
-            len++;
-        } else if ('A' <= *hex && *hex <= 'F') {
-            result <<= 4;
-            result += (*hex) - 'A' + 10;
-            len++;
-        } else {
-            return "";
-        }
-        hex++;
-    }
-    return 0 < len ? std::to_string(result) : "";
-}
-
-std::string bin2dec(const char* bin)
-{
-    int result = 0;
-    int len = 0;
-    while (*bin) {
-        if ('0' == *bin || '1' == *bin) {
-            result <<= 1;
-            result += (*bin) - '0';
-            len++;
-        } else {
-            return "";
-        }
-        bin++;
-    }
-    return 0 < len ? std::to_string(result) : "";
-}
